requirements.txt
albumentations
einops
extra_streamlit_components
h5py
joblib
kornia
loguru
matplotlib
numpy
opencv_python
pandas
Pillow
Pillow
pytorch_lightning
streamlit
streamlit_option_menu
torch
tqdm
yacs
